Add a "reset victories" option to the main manu
The player can now reset the number of victories to 0. This does not
affect achievements. It is in preperation of more widespread use of
unlocking content. (Players might want to reset their profile and start
unlocking the locked content again -- or for the first time, if they
have played an older version of Kerkerkruip.)