Chain golem and swarm of daggers changes
* Changed the dex check of the pierce ability from 10 + 2 x greatest
power to 8 + concentration + 2 x greatest power.
* Ducking under the chains is now a "first carry out attacking" rather
than a "check attacking" rule. This stuff only happens when the attack
has commenced, right? This also solves a bug where using the piercing
ability on the golem would require the player to duck under the chains.
* Ducking under the chains is nwo also necessary when you have no
concentration. That this was not the case was a bug! (At one point, the
chains only made you lose concentration, so the restriction made sense,
but now that the chains deal damage as well, it makes no sense.)