Bodice of breathlessness
The bodice of body (used to be shirt of muscles), now has a possible
cursed version. This makes the symmetry between the crown of the empire,
the diadem of the priestess, and the bodice of body complete. All of
them give a +2 bonus, and all have a cursed version which also gives a
faculty bonus, but has some nasty side effect.

The bodice of breathlessness grants a +2 body bonus as normal... but
this is lowered by 2 for every points of concentration, as the bodice's
lace tightens. At 3 concentration, you'll have a -4 body penalty.