Magical guides
Added a new type of item, the magical guide. Reading a magical guide
identifies a couple of scroll types. There are currently four magical
guides:

1. Introduction to Necromancy (death, summoning, ghoulification)
2. Battle Magic for Beginners (the blade, protection)
3. Handbook of Subtle Magics (skill, shadows, teleportation)
4. Miscellanea Magica (three random scroll types)

Because I'm using the scroll names, there are currently two problems:
"the-blade" and "shdoaws-name" don't look like they should to the
player. I still need to fix that.