Bug fix: smoke demon remembered when gone
One of those subtle cases: remember showed the last seen location of all
alive seen person. Normally, that would mean that no off-stage persons
are shown, because someone who is seen (and thus has been on-stage) will
only be put off-stage when dead. But the smoke demon is an exception to
this rule, and was therefore sometimes remembered to be in a place where
the player had already seen him disappear.