Testobjects
We have a new "not for release" section in Dungeon Generation. We can
now designate any room or thing as a testobject, which will:

- for a room, almost ensure that it will be placed by the normal dungeon
generation rules (probability of it not being placed is VERY low)
- for a monster with a level, ensure that it will be placed by the
normal monster placement rules
- for anything else, ensure that it will be put somewhere in the
dungeon.

Of course, "purloin" and "gonear" are often good enough for testing, but
I've had situation where I would like the normal procedures to place
stuff.