Groups and powers
(Needs a bit more testing, but I need to go away now for most of the
rest of the day.) Removes fanatics-of-Aite-package, and ensures that
powers are granted in the right way.

I had a LOT of trouble because we defined grouper wrongly:

-Definition: a person is a grouper if it relates to a person by the
accompaniment relation.
+Definition: a person is a grouper if a person relates to it by the
accompaniment relation.

I'm not sure I understand the logic of that syntax!!