Check whether power is granted
Added a check so that powers are not granted if the player already has
them. (Plonking the fanatics of Aite would grant the power three times.
Not sure whether that would also happen without cheating commands --
say, if you kill them all with one fragmentation grenade -- but adding a
check certainly won't hurt.)