Implemented Sul (except for intervention)
All favours granted by Sul are now implemented, including angelic
radiance.

Not really related, but I've also made it slightly more likely that the
crown of the empire is cursed, and quite a bit more likely that the
diadem of the priestess is (because its curse is so much fun!).