Arena of the Gods #86 - First pass
Arena of the Gods, which are extensions to Kerkerkruip Religion and
Kerkerkruip Locations. Still has test values, since we lack Chosen Ones
(see #94). Divine Intervention still has to be implemented. For now, the
god grants 2 divine favour. I'm actually thinking about making a special
weapon / armour or other type of item for each of the gods as an
alternative gift (a "branded weapon" as in DC:SS). I'm considering
adding some monsteravatars myself.