Two imp changes
* Fix a bug with the teleportation beacon. (Now really.)
* Made it so that the imp cannot be defeated if he can teleport away --
he'll always teleport. This means that you'll have to find the Lair of
the Imp (which can always be done) of stop it from teleporting in some
other way. This change had to be made because otherwise the optimal
strategy for dealing with the imp is the BORING one of waiting in the
Entrance Hall until it shows up often enough to get killed.