NEW: The mausoleum
The mausoleum is a treasure room that is sometimes placed unconnected,
and sometimes connected to the dungeon. The tomb contains at least one
epic item ... but it is guarded by a super-undead, or at least two
undead (who can be super-undead). Opening the tomb in combat takes a lot
of time, so you can't just run in and grab the stuff without defeating
the enemies.