Imp only appears in combat
Why? We don't want the optimal strategy to hang around somewhere and
wait for the imp to appear, hit it, wait until returns, and so on. So I
changed the imp to always teleport away when it was attacked (if
possible); this means that it can only be defeated in its own Lair or
when the dimensional anchor is present. However, this creates a new
problem: whenever the imp appears, the player should type "a imp". That
will make it disappear. Until it reappears a couple of turns later ...
boring!

So now the imp will only appear in combats, where attacking it is not a
free action and might not be worth it.