Solves issue #142
Added arena-faction. When entering the arena's, the faction of the
player becomes arena-faction. If scroll of ghoulification is read in the
Arena of the Gods, it reports differently and the faction of the player
is reverted to the Arena-faction. After exiting the Arena, it is
returned to the original faction.