First pass at Automatos
Automatos is a new level 4 creature, and a champion of Nomos. He is vert
strong, but follows a completely determined pattern of movements -- for
instance, he will attack, concentrate, block, concentrate, attack,
dodge, block, concentrate, and then starts the process over from the
beginning.

The AI basics are in place, but need to be tweaked. Prose must be
written, a power must be written, and so on.