Shield of fanaticism
The defender of Aite didn't have a shield! And in fact never blocked. I
still need to change the prose, mind you.
We now also have the shield of fanaticism, which is very strong when all
disciples of Aite are present. Perhaps we need to slightly tweak the
defense rating of the defender.