Put xorshift in charge of automated tests
This should eliminate most differences in
test runs between gargoyle and the IDE.

As responsible testers, we should manually update the
random-seed entry of test queue rows. The way I've been doing this is like so:

1. Test a single set until all tests pass. Then increment the default random-seed.
Repeat until the single set passes twice in a row on two different seeds.

2. Run all tests. If any test fails, go to step 1.

I've just done a couple iterations of this and modified a few tests
so they pass consistently.