Made demon of rage waking consistent with angel of compassion
In a lot of places I have added a check to see if someone is
"sleeping in this world" before waking them up. In a big way this is
arbitrary at the moment because we have no dreams in which we interact
with people sleeping in the dungeon. But for now I've decided to follow
this rule of thumb:

If an effect wakes somebody up, it should first check if that person is
sleeping in this world, IF that effect could target someone chosen by the
player.

If the effect could target someone not chosen by the player, but
the target shares a world with the player, I'd lean towards checking
first, but it's more of a judgment call.

Finally, it is always dangerous to wake the player up, because it can
interrupt the normal course of dreams. If the target of
a waking-up effect could potentially be the player, definitely check first.