Clarified nomos counter
I discovered that the Nomos counter was counting
down an extra time if Nomos intervened on the player's
turn. I fixed it so the player still gets a shorter
countdown if that happens, but it's accurately reported.

I also added Nomos attack details to the status info.

Other changes:

fixed Angel of Mercy shrinking message when dead.

Added wait-canceling for Hand of Glory melting and Hand of Vecna
betrayal.