Progress on a cross-platform "make test" target
Everything except the interpreter works now -
on my system I have manually installed a cheapglk-based
version of glulxe and symlinked it to the "dumb-git"
interpreter that the test runner tries to use. This will
not work on other Mac machines but everything else
in the test runner seems to work on Mac now.

 Referenced in #399