More angel of Mercy tweaks
I increased the angel's defence to 11 -
maybe this is too much - it could be overpowered
and possibly lead to drawn-out battles.
Perhaps it should get a slightly higher damage
die for its natural weapon instead.

There also seems to be some freezing issue -
a fight test between the tentacle
and the angel of mercy hung up mysteriously