Fix some obvious pile of body parts glitches
The Hand of Vecna was never actually
provided, even when it said it was.

The Hand of Glory can be given by the Angel of Mercy,
so we needed to check if it was off-stage before
finding it in the pile.

It might be better to give the hand of glory a "melted" property
to keep it from being provided twice.

Also attempting to fix the mercy-radiation in dreams
glitch. It would be better to write a test for this.