Fixed some tests, tried to make a few tests faster
I realized that I should have been increasing the achievement
target for tests, not lowering it. A higher achievement
target means a lower chance of false negatives (with a higher
chance of false positives). We're getting too many
false negatives, so this seemed like a good idea.

I also reduced the maximum attempts of several tests - because
of the way the achievement target is used, this should affect
the rates of false positives and negatives equally, and should
make the tests shorter.

Phantasmagoria tests were moved out of Simple Tests just because
they're slow

I changed the blessed grenade test to avoid using regular
expressions. Hopefully that makes it faster, but it's still pretty slow