Fix some bugs introduced during refactoring
There were some issues regarding whether a room
is "secretly placed" or "normally placed."

I removed the code that automatically marks a room
as secretly placed - that is now done only
with an explicit call to the phrase.

There is a warning, though, when attempting
to place a room "next to" a secret room.

I also changed the name of "mark (room) as normally placed"
because it doesn't agree with the "normally placed" adjective.