Got rid of the "standard measure" object
It's no longer needed now that there is a
"size number" phrase.

It was only used in two places. One of them was when reading a scroll
of summoning as a reaction. The size of the creature summoned
affects the spirit target for the attacker to hit their
intended target and not the creature summoned.

The way it looks to me, the original code had a lower
target for larger creatures summoned. This seems wrong,
so I reversed it.

If anything is not right with this, we should revisit it.
Making a test might not be a bad idea.