Add a simple test and fix #416
I've added a new rule,
the "fix uncleared Nomos bonus rule"
that should fire whenever the Nomos bonus
forces someone to attack but the attack fails
(i.e. never reaches the aftereffects phase)

As far as I know, the only way this can happen
is if the current weapon of the Nomos attacker
is an unloaded crossbow. But we should
be on the lookout for other possibilities. I
still feel that the Nomos attack mechanic is
a bit fragile.

The new rule punishes the Nomos attacker
for failing to attack, destroying their
weapon and doing 5 damage. I don't know of
any way this could happen when they have
a natural weapon, but that case is covered
anyway by doing damage without destroying
the weapon.