Refactor concentration losing
Rewrite the code for concentration losing to allow it to be called from
outside combat, ensuring that skills and items that allow you to keep
concentrated also function for these effects. This particular patch
makes it work for the mindslug's mindblast and the tentacle's shaking,
but a few other places still have to be tackled.