Cosmetic sweat lodge changes
The room description was very long in the game; I've shortened it, even
though we sacrifice some flavour. (The presence of people you could not
interact with was, I think, a bit strange, especially for new players.)
I made the sweat lodge and the hall of the raging banshees rarer. These
are special places.